错误处理移动
